Lookup User Ids (#30)
Because user screen names can change over time, it would be more accurate to construct retweet graphs using user ids than screen names. However the way the tweets were collected, if it is a retweet, we have access to the retweeted user's screen name but not their id. So this PR introduces a script to lookup those user ids.

FYI: Some user ids are not found due to account termination or suspension.